+++ api/src/graphql/ik/attempts.sdl.ts
@@ -3,27 +3,10 @@ export const schema = gql`
     stepId: String!
     data: JSON!
   }
-
-  type StepProgress {
-    stepId: String!
-    solved: Boolean!
-    stepSortWeight: Int!
-  }
-
-  type StepsByPuzzleIdResponse {
-    puzzle: Puzzle!
-    stepProgress: [StepProgress!]!
-  }
-
   type MakeAttemptResponse {
     success: Boolean!
     finalStep: Boolean
     message: String
-  }
-
-  type Query {
-    getStepsByPuzzleId(id: String!): StepsByPuzzleIdResponse @skipAuth
-    optionalStep(id: String, puzzleId: String, stepNum: Int): Step @requireAuth
   }
 
   type Mutation {
