+++ api/src/services/ik/attempts/attempts.ts
@@ -1,15 +1,10 @@
-import { PUZZLE_COOKIE_NAME } from '@infinity-keys/constants'
-import cookie from 'cookie'
-import type { MutationResolvers, QueryResolvers, StepType } from 'types/graphql'
+import type { MutationResolvers, StepType } from 'types/graphql'
 import { z } from 'zod'
 
-import { context, ForbiddenError } from '@redwoodjs/graphql-server'
+import { context } from '@redwoodjs/graphql-server'
 
-import { isAuthenticated } from 'src/lib/auth'
 import { db } from 'src/lib/db'
-import { decryptCookie } from 'src/lib/encoding/encoding'
 import { createSolve } from 'src/services/solves/solves'
-import { step } from 'src/services/steps/steps'
 import { createUserReward } from 'src/services/userRewards/userRewards'
 
 import { checkNft } from '../minter/check-nft'
@@ -214,38 +209,4 @@ export const makeAttempt: MutationResolvers['makeAttempt'] = async ({
     console.log(e)
     return { success: false }
   }
-}
-
-export const optionalStep: QueryResolvers['optionalStep'] = async (
-  { id, puzzleId, stepNum },
-  { context }
-) => {
-  // users don't need specific step data on the puzzle landing page
-  if (!id) return
-
-  // unauthenticated users can't see specific step data
-  if (!isAuthenticated()) {
-    throw new ForbiddenError('must sign in')
-  }
-
-  // authenticated users should be allowed to view first steps
-  if (stepNum === 1) return step({ id })
-
-  const puzzlesCompletedCypherText = cookie.parse(context.event.headers.cookie)[
-    PUZZLE_COOKIE_NAME
-  ]
-
-  const puzzlesCompleted = decryptCookie(puzzlesCompletedCypherText)
-
-  // get the number of completed steps for user on this puzzle. they should
-  // be allowed to view previous steps and the current step they are on
-  const visibleSteps = puzzlesCompleted?.puzzles[puzzleId]?.steps.length + 1
-
-  // ensure they've solved the correct number of previous steps for the step
-  // they are trying to view
-  if (stepNum > visibleSteps || !visibleSteps) {
-    throw new ForbiddenError('Step currently not viewable.')
-  }
-
-  return step({ id })
 }
